modularize gatsby-theme-series gatsby-node
